Move async test helpers to rollbar.lib.async
